Split DochiViewModel into focused modules (259 lines from 503)
Extract remaining responsibilities from DochiViewModel:
- ConversationManager: conversation history CRUD and recent summaries (88 lines)
- DochiViewModel+Callbacks: service callback wiring and Combine bindings (164 lines)
- TTSSanitizer: markdown-to-speech text sanitization utility (29 lines)

Module breakdown (total 1113 lines, was 1062 in single file):
- DochiViewModel.swift: 259 lines — coordinator, state, public API
- DochiViewModel+Callbacks.swift: 164 lines — service wiring
- SessionManager.swift: 175 lines — session lifecycle
- ToolExecutor.swift: 132 lines — tool execution loop
- ContextAnalyzer.swift: 266 lines — conversation analysis, memory
- ConversationManager.swift: 88 lines — conversation history
- TTSSanitizer.swift: 29 lines — TTS text cleanup

Closes #1
